Complete dom updates and fix file structure
