+++ src/classes/IngredientClass.js
@@ -7,4 +7,3 @@ class Ingredient {
 };
 
 export default Ingredient;
-
