+++ src/css/index.scss
@@ -1,0 +1,1 @@
+@import './_reset.scss';
