+++ src/styles.css
@@ -1,4 +1,3 @@
-
 * {
   font-family: 'Palanquin', sans-serif;
 }
